Auth refactoring. Adding access to auth data from outer world

+++ src/app/services/auth/auth.service.ts
@@ -2,15 +2,13 @@ import { Injectable } from '@angular/core';
 import { LocalStorageService } from '../local-storage/local-storage.service';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
+import { ApiResponseData } from '../api.types';
+import { LoginData } from './auth.types';
 
 
 
 const AUTH_STORAGE_KEY = 'auth';
 
-interface AuthData {
-  login: string,
-  password: string,
-}
 
 @Injectable({
   providedIn: 'root'
@@ -18,7 +16,7 @@ interface AuthData {
 export class AuthService {
 
   private _loggedIn = false;
-  private _authData: AuthData;
+  private _loginData: LoginData;
 
   constructor(private http: HttpClient, 
               private storage: LocalStorageService) { }
@@ -28,33 +26,45 @@ export class AuthService {
     this._loggedIn = this.updateCredentials();
   }
 
-  public login(login: string, password: string): Promise<boolean> {
+  public login(login: string = '', password: string = ''): Promise<ApiResponseData> {
+    login.trim(); password.trim();
 
-    return new Promise<boolean>((resolve) => {
+    return new Promise<ApiResponseData>((resolve) => {
       
-      new Observable<{success: boolean}>( s => {
-        // HTTP Request
-        s.next({success: login == 'demo' && password == 'demo'});
+      if(!login || !password) {
+        resolve({
+          success: false,
+          message: 'Заполните все поля'
+        });
+        return;
+      }
 
-      } ).subscribe(res => {
+      this.http.get<ApiResponseData>('https://api.pricer.versla.ru/auth/login', {
+        params: {
+          login: login,
+          password: password,
+        }
+      }).subscribe(res => {
         
         if(res.success) {
 
-          this._authData = {
+          this._loginData = {
             login: login,
             password: password,
           };
 
-          this.storage.writeData(AUTH_STORAGE_KEY, this._authData);
+          this.storage.writeData(AUTH_STORAGE_KEY, this._loginData);
 
         } else {
           this.storage.delete(AUTH_STORAGE_KEY);
-          this._authData = null;
+          this._loginData = null;
         }
 
         this._loggedIn = res.success;
-        resolve(res.success);
+        resolve(res);
         
+      }, res => {
+        resolve(res.error);
       });
 
     });
@@ -64,7 +74,9 @@ export class AuthService {
     return this._loggedIn;
   }
 
-
+  get loginData(): LoginData {
+    return this._loginData;
+  }
 
 
   private updateCredentials(): boolean {
@@ -72,7 +84,7 @@ export class AuthService {
     if(!this.storage.exists(AUTH_STORAGE_KEY))
       return false;
 
-    this._authData = this.storage.readData(AUTH_STORAGE_KEY);
+    this._loginData = this.storage.readData(AUTH_STORAGE_KEY);
 
     return true;
   }
